objects and shadows working just fine, some issues with the cylinder

+++ parser/is_str.c
@@ -6,7 +6,7 @@
 /*   By: hlimouni <hlimouni@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/19 17:01:03 by hlimouni          #+#    #+#             */
-/*   Updated: 2021/01/21 14:38:25 by hlimouni         ###   ########.fr       */
+/*   Updated: 2021/01/22 08:27:59 by hlimouni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		is_str_uint(char *str)
 {
-	int		i;
+	size_t		i;
 
 	if (!str)
 		return (0);
